+++ src/site/changelog/types.rs
@@ -1,0 +1,36 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct ReleasesApiResponse {
+    pub url: String,
+    pub assets_url: String,
+    pub html_url: String,
+    pub id: i64,
+    pub tag_name: String,
+    pub target_commitish: String,
+    pub name: String,
+    pub draft: bool,
+    pub prerelease: bool,
+    pub created_at: String,
+    pub published_at: String,
+    pub assets: Vec<Asset>,
+    pub tarball_url: String,
+    pub zipball_url: String,
+    pub body: Option<String>,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Asset {
+    pub url: String,
+    pub id: i64,
+    pub node_id: String,
+    pub name: String,
+    pub label: String,
+    pub content_type: String,
+    pub state: String,
+    pub size: i64,
+    pub download_count: i64,
+    pub created_at: String,
+    pub updated_at: String,
+    pub browser_download_url: String,
+}
